Ajout des clients Feign et mise à jour des microservices

+++ matiere/src/main/java/com/etec/matiere/entity/Matiere.java
@@ -16,6 +16,6 @@ public class Matiere {
     private Long id;
 
     private String nom;
-    private double coefficient;
+    private Integer coefficient;
 
 }
